+++ RELEASE-NOTES.md
@@ -10,3 +10,4 @@
 * #13 Support Method injection
 * #15 Remove injection of static fields #15
 * #16 Configurable default injection scope
+* #17 Configurable annotation types and provider implementation